refactor: move list model to list package
